Simplify support for smali injections for odex builds
Until now smali injections have been performed using the squisher
which is run to modify the final ota package. The problem with this
is that odex'd and deodex'd builds need separate code paths to
handle baksmali/smali and it seems like it would be better to just
go ahead and inject the smali before dexopt. This adds support for
running a custom script similar to squisher prior to creating
system.img. This allows smali to be injected once using the same
code for both odex/deodex builds.

Change-Id: Ib5438dc727af97ff4cd37ed3a6b32feeed2aad23